document-search-airflow/pyproject.toml
[project]
name = "document-search-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "awscli-local[ver1] (>=0.22.0,<0.23.0)",
    "invoke (>=2.2.0,<3.0.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pypdf (>=5.7.0,<6.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
